The DoxygenPlugin for Trac was fully compatible with [http://www.stack.nl/~dimitri/doxygen/ Dowygen] only for versions prior to 1.8.0, since the [http://www.stack.nl/~dimitri/doxygen/manual/changelog.html#log_1_8_0 search tools] have been completly redesigned since.
We propose here a new version that rely completly on the new {{{searchdata.xml}}} fileindexing the whole documentation managed by Doxygen. This allows to:

 * have a direct link to the documentation of a function and the like in a Doxygen file when using TracLink syntax {{{[doxygen:foo] and [doxygen:project/foo]}}};

 * discard the code searching for a Doxygen file, since it was not compatible with the CREATE_SUBDIR option, was not realiabale for several files with the same name in different directories, and causes security problems.

As a consequence, the options "ext" et "source_ext" are now useless, and conversely there is now the "searchdata_file" option, indicating the name of the index file computed by Doxygen, corresponding to its "SEARCHDATA_FILE" option, and with the same default value ("searchdata.xml"). Setting the SEARCHENGINE option to YES in Doxygen is mandatory for having the TracLinks .

Note also that the "encoding" option is now "utf-8" by default.

The integration of the two search engines (Trac and Doxygen) is still working,
however there is presently no more ranking.